:tada: Add support for batching
We now have a 'messages' endpoint for http, socketio, and osc that will
call all of the asserts and retracts to the database in order before
executing them all.

In addition, we can run multiple services from the same machine, and
they will just pick different ports. This is great for testing, which
now happens in paralell.